+++ libs/blockscout-client/crate/src/models/extended_revert_reason_as_map.rs
@@ -13,13 +13,13 @@
 */
 
 use crate::models;
+use rust_decimal::Decimal;
 use serde::{Deserialize, Serialize};
 use serde_aux::prelude::*;
 
 #[derive(Clone, Default, Debug, PartialEq, Serialize, Deserialize, derive_new::new)]
 pub struct ExtendedRevertReasonAsMap {
     pub raw: String,
-    #[serde(deserialize_with = "deserialize_number_from_string")]
-    pub code: i32,
+    pub code: Decimal,
     pub message: String,
 }
