+++ libs/blockscout-client/crate/src/models/raw_trace.rs
@@ -13,14 +13,14 @@
 */
 
 use crate::models;
+use rust_decimal::Decimal;
 use serde::{Deserialize, Serialize};
 use serde_aux::prelude::*;
 
 #[derive(Clone, Default, Debug, PartialEq, Serialize, Deserialize, derive_new::new)]
 pub struct RawTrace {
     pub action: models::RawTraceAction,
-    #[serde(deserialize_with = "deserialize_number_from_string")]
-    pub subtraces: i32,
+    pub subtraces: Decimal,
     pub trace_address: Vec<i32>,
     pub r#type: String,
     #[serde(default)]
